aws/lambda_handler.py
import json
import boto3

sagemaker_runtime = boto3.client('sagemaker-runtime')

def lambda_handler(event, context):
    # Get current time and recent train history
    current_time = event.get('timestamp')
    recent_history = get_recent_train_history()
    
    # Process features
    features = process_features(current_time, recent_history)
    
    # Get predictions from SageMaker endpoint
    response = sagemaker_runtime.invoke_endpoint(
        EndpointName='train-probability-endpoint',
        ContentType='application/json',
        Body=json.dumps(features)
    )
    
    # Format and return predictions
    predictions = format_predictions(response)
    return {
        'statusCode': 200,
        'body': json.dumps(predictions)
    }

def get_recent_train_history():
    """Retrieve recent train crossing history"""
    # Implementation details would go here
    pass

def process_features(current_time, recent_history):
    """Process input data into features"""
    # Implementation details would go here
    pass

def format_predictions(response):
    """Format the model predictions into the desired response structure"""
    # Implementation details would go here
    pass 