pyproject.toml
[tool.poetry]
name = "rail-ml-pipeline"
version = "0.1.0"
description = ""
authors = ["Scott Alan Turner <turner655+github@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}] 

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.3"
numpy = "^2.1.3"
scikit-learn = "^1.5.2"
seaborn = "^0.13.2"
matplotlib = "^3.9.2"
prophet = "^1.1.6"
statsmodels = "^0.14.4"
plotly = "^5.24.1"
xgboost = "^2.1.2"
boto = "^2.49.0"


[tool.poetry.group.dev.dependencies]
jupyter = "^1.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
